simplify approach dispatch

+++ group_by_hashtable.cuh
@@ -222,7 +222,7 @@ __global__ void kernel_write_out_group_ht(
 }
 
 static inline bool approach_hashtable_available(
-    int group_bits, int row_count, int grid_dim, int block_dim,
+    int group_bits, size_t row_count, int grid_dim, int block_dim,
     int stream_count)
 {
     if (!grid_dim || !block_dim) return false;
