+++ group_by_shared_mem_hashtable.cuh
@@ -16,7 +16,7 @@ struct shared_mem_ht_entry {
 GROUP_BY_HASHTABLE_FORWARD_REQUIREMENTS(group_by_shared_mem_hashtable)
 
 static inline bool approach_shared_mem_hashtable_available(
-    int group_bits, int row_count, int grid_dim, int block_dim,
+    int group_bits, size_t row_count, int grid_dim, int block_dim,
     int stream_count)
 {
     if (!grid_dim || !block_dim) return false;
