+++ group_by_shared_mem_perfect_hashtable.cuh
@@ -240,7 +240,7 @@ static inline void group_by_shared_mem_perfect_hashtable_fin()
 }
 
 static inline bool approach_shared_mem_perfect_hashtable_available(
-    int group_bits, int row_count, int grid_dim, int block_dim,
+    int group_bits, size_t row_count, int grid_dim, int block_dim,
     int stream_count)
 {
     if (!grid_dim || !block_dim) return false;
